Move the tests out of src and into a top-level tests dir
Ten test files sat next to the source they cover, so the PR diff interleaved
them with the code and there was no way to read one without the other.

tests/ mirrors src/ — lib, tools, and the two stubs under helpers. Nothing
changed but the paths, the imports that follow them, and the two configs that
point at them. Git records all twelve as renames.
