+++ tests/helpers/raycast-stub.ts
