+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "$schema": "https://json.schemastore.org/tsconfig",
     "display": "Node 16",
-    "include": ["src/**/*"],
+    "include": ["src/**/*", "tests/**/*"],
     "compilerOptions": {
       "lib": ["es2021"],
       "module": "commonjs",
